refactor: extract 'KeyUtils' to :common:android
Simple KeyEvent digit helpers with no project dependencies. Lives in
:common:android since it uses android.view.KeyEvent.
