+++ common/android/src/test/java/com/ichi2/anki/common/utils/android/KeyUtilsTest.kt
@@ -14,44 +14,40 @@
  *  this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.ichi2.utils
+package com.ichi2.anki.common.utils.android
 
 import android.view.KeyEvent
 import android.view.KeyEvent.ACTION_UP
 import android.view.KeyEvent.KEYCODE_0
+import android.view.KeyEvent.KEYCODE_5
 import android.view.KeyEvent.KEYCODE_9
+import android.view.KeyEvent.KEYCODE_BACK
 import android.view.KeyEvent.KEYCODE_ENDCALL
 import android.view.KeyEvent.KEYCODE_STAR
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.ichi2.anki.EmptyApplicationCategory
-import com.ichi2.testutils.EmptyApplication
-import com.ichi2.utils.KeyUtils.getDigit
-import com.ichi2.utils.KeyUtils.isDigit
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
-import org.junit.Assert.assertTrue
 import org.junit.Test
-import org.junit.experimental.categories.Category
 import org.junit.runner.RunWith
-import org.robolectric.annotation.Config
+import kotlin.test.assertEquals
+import kotlin.test.assertNull
 
+/** Tests for [KeyEvent.digit]. */
 @RunWith(AndroidJUnit4::class)
-@Config(application = EmptyApplication::class)
-@Category(EmptyApplicationCategory::class)
 class KeyUtilsTest {
     @Test
-    fun testIsDigit() {
-        assertTrue(isDigit(KeyEvent(ACTION_UP, KEYCODE_0)))
-        assertTrue(isDigit(KeyEvent(ACTION_UP, KEYCODE_9)))
-        assertFalse(isDigit(KeyEvent(ACTION_UP, KEYCODE_STAR)))
-        assertFalse(isDigit(KeyEvent(ACTION_UP, KEYCODE_ENDCALL)))
+    fun `digit returns the digit for 0-9 keys`() {
+        assertEquals(0, KeyEvent(ACTION_UP, KEYCODE_0).digit)
+        assertEquals(5, KeyEvent(ACTION_UP, KEYCODE_5).digit)
+        assertEquals(9, KeyEvent(ACTION_UP, KEYCODE_9).digit)
     }
 
     @Test
-    fun testGetDigit() {
-        assertEquals(0, getDigit(KeyEvent(ACTION_UP, KEYCODE_0)))
-        assertEquals(9, getDigit(KeyEvent(ACTION_UP, KEYCODE_9)))
-        assertEquals(null, getDigit(KeyEvent(ACTION_UP, KEYCODE_STAR)))
-        assertEquals(null, getDigit(KeyEvent(ACTION_UP, KEYCODE_ENDCALL)))
+    fun `digit returns null for non-digit printable keys`() {
+        assertNull(KeyEvent(ACTION_UP, KEYCODE_STAR).digit)
+    }
+
+    @Test
+    fun `digit returns null for non-printable keys`() {
+        assertNull(KeyEvent(ACTION_UP, KEYCODE_ENDCALL).digit)
+        assertNull(KeyEvent(ACTION_UP, KEYCODE_BACK).digit)
     }
 }
